Publicise final fields consituting Record fields.
The values that are read from the CDF file and stored directly as
final fields in the corresponding Record subclasses are now available
as public final fields in the corresponding classes.  Since they
are final this is pretty harmless, and saves a lot of tedious and
perhaps error-prone writing of accessor methods.