RecordFactory now creates records by hand, not using reflection.
As well as eliminating reflection which is in general a questionable thing
to use, it allows custom construction of Record instances (they don't
all need to have the same constructor signature).  This in turn provides
a hook for per-instance variation of RecordFactory behaviour, which
we will need for coping with differnent CDF versions.