Removing parameterised types from most DataType method signatures.
Suppress the warnings with paired non-parameterised methods,
and hide the parameterised ones.  Since in most cases you won't
know what DataType you're using, it's a pain to have to deal
with the checking.
So why make DataType a parameterised type at all?
It's mostly for compile-time checks within the DataType class itself,
and to make it clear what's going on there.  It also provides
documentation within the interface (javadocs) of what types
the DataTypes use.  That may be good or, if it has to change, bad.