Make use of GDR LeapSecondLastUpdated field for warnings/errors
The GDR LeapSecondLastUpdated field introduced in CDF v3.6 is now
recorded and used when making sense of TIME_TT2000 data values.
If the library leap second table is out of date, a warning is issued.
If the time values are known to have leap seconds applied invalidly,
an error is thrown or a severe log message is issued.  This behaviour
follows that of the official CDF library (email to MBT from Robert
Candey, 20 Apr 2015).

Implementing this requires DataType instances to be sensitive in at
least some cases to the details of the file they are used with,
so required a little bit of alteration to DataType and some other
classes.  DataType instances are no longer always static.  But I
don't think it should break anything.