Use NASA CDF v3.6.0.4 in tests
Use the cdfjava.jar file from v3.6.0.4 of NASA's CDF library.
This fixes a bug present in v3.6.0.3; it lacked the 2015 leap second.
This change means we can simplify the test invocations somewhat.
Take the opportunity to package the cdfjava JAR file along with
this package, since it isn't very large, thus removing any
dependency on an external CDF library installation for tests.

The file nasa/cdfjava_3.6.0.4.jar was downloaded from
http://cdaweb.gsfc.nasa.gov/pub/software/cdf/dist/cdf36_0/linux/cdfjava.jar
on 09-Sep-2015.