Fix a bug which prevented access to large (>2Gb) files
In one place I was using 32-bit integers on the right hand side
of an assignment to a 64-bit integer.  This truncated the value
before it got converted to 64 bits :-(.
Declared one of the variables that participates in the arithmetic
as long instead of int and it's OK.

The large-file access had never been tested because I'd never
been able to find a multi-gigabyte CDF file.
Jeremy Faden kindly contacted me and and pointed me at
http://themis.ssl.berkeley.edu/data/rbsp/rbspa/l1/mscb1/2015/
which has CDF files (though they're not particularly interesting
ones for my purposes) up to 15Gb.  Testing on those easily
revealed the bug.