Makefile option to reduce noise when running JDK 8 javadoc
In JDK v8 javadoc issues a lot of warnings that were not present
in earlier versions.  Some are useful, but I don't plan to write
@throws clauses for all thrown exceptions, and each of those
generates a warning.  Add a line (javadoc flag -Xdoclint:all,-missing)
in the makefile that you can uncomment if you're using java 8
and the warnings get boring.
Unfortunately this flag causes an error in earlier javadoc versions,
so you can't just leave it in unconditionally.