Avoid timezone-dependence of TIME_TT2000 value formatting
Previously, the calendar handling used to format TIME_TT2000 date
values used a GregorianCalendar instance with the system-default
time zone, so the value you got was dependent on the timezone
you were running in (could be out by up to a second within a few
hours of a leap second).  Fix it.