Use external leap seconds config file for testing
For othertest, instead of relying on the leap seconds table internal
to the NASA CDF library, the test now uses a CDFLeapSeconds.txt
file that can be downloaded from the CDF web site.  This makes it
easier to update the leap seconds information for tests.

The makefile provides a target for updating this leap seconds file
when a new one is available, but the file is also checked into
the repository for convenience.