Replace CDF test files with uncopyrighted versions
The testcases now use test files from the CDF V3.6.5 distribution
(probably, but see below).
At this commit the test files themselves are updated, and the unit
tests have been adjusted correspondingly, since the data content of
these test files is similar, but not identical, to their incarnations
from the CDF V3.4.1 distribution that were used before.

The point of this change is that V3.4.1 testfiles contained a
copyright statement that was incompatible with the GPL,
so the package including test files could not be packaged
for Debian.  The new test files are apparently harmless.

The actual replacement test files included in this commit were
supplied on request by the NASA CDF office (Michael Liu via
Robert Candey at GSFC), using code apparently destined for
CDF V3.6.5.  However, at the time of this commit the V3.6.5
release has not been made, so it's possible that details of
these files will differ from those actually packaged with
CDF V3.6.5 when it is released.