Explicitly compile for Java 1.5
Add a javac flag -target 1.5 in the makefile.  The source code is
currently 1.5 compatible, so I might as well make sure that the
jar file uses the java 1.5 class file binary format for maximum
backwards compatibility (though I doubt if anybody is using it
with Java 1.5 still).

If future changes to the source code require later JRE platforms,
that compiler flag can be removed or adjusted.