fix bug interpreting GDR leapSecondLastUpdated field
Fix a bug that identified the wrong TtScaler from the date supplied
in the leapSecondLastUpdated field of the Global Data Record.
Because it used leap seconds in mapping the YYYYMMDD value to epoch,
and hence arrived a few seconds earlier than the nominal date,
it generally picked up the scaler one or two older than it should do.
That led to reports that the library leap second table was out of date
with respect to the data, which in turn yielded RuntimeExceptions
("CDF TIME_TT2000 date formatting failed - library leap second table
known to be out of date with respect to data") when formatting TT2000
values coming after the CDF file's leapSecondLastUpdated value.

Actual TT2000 formatting calculations were not affected.